+++ src/features/secret-vault/index.ts
@@ -1,0 +1,2 @@
+export * from './SecretVaultPage'
+export * from './useSecretVault'